refactor: migrate Landing
